Separated map initialization and route rendering into individual files
